requirements.txt
torch
numpy
matplotlib
onnx
onnx-simplifier
tqdm
visualtorch